UploadService/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.35.96
botocore==1.35.96
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
jmespath==1.0.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
